shared/imgproc: updated interface (using ulong to possibly huge numbers)
git-svn-id: https://scm.citationtech.net/trunk/shared/imgproc/imgproc@9470 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ imgproc/rastermask/quadtree.hpp
@@ -69,10 +69,10 @@ public :
     bool onBoundary( int x, int y ) const;
 
     /** return mask size (number of white pixels) */
-    uint count() const { return count_; }
+    ulong count() const { return count_; }
 
     /** return total number of pixels */
-    uint capacity() const { return sizeX_ * sizeY_; }
+    ulong capacity() const { return ulong(sizeX_) * ulong(sizeY_); }
 
     /** test mask for emptiness */
     bool empty() const { return count_ == 0; }
